Fixed time stepping.  Physics and do_timer now cycle at 60Hz, regardless of actual frame rate.

+++ src/script.h
@@ -23,7 +23,7 @@ void add_argument(int, const char *);
 
 int  do_click_script(int, int);
 int  do_point_script(int, int);
-int  do_timer_script(int);
+int  do_timer_script(float);
 int  do_frame_script(void);
 int  do_keyboard_script(int, int);
 int  do_joystick_script(int, int, int);
